feat: add apiKeyRepairEnabled AppSync configuration option
Set apiKeyRepairEnabled to `true` to automatically create a new, unique AppSync API key CFN object if the plugin-maintained CFN object is associated with an AppSync API key id that no longer exists.  This can happen on API key expiration or if an API key is manually deleted from the AWS console.  If this condition is present, the deployment will fail without this option enabled because CFN will try to look up the deleted API key by ID and fail.  This option is only relevant for authenticationType === API_KEY

* Modify mergeCustomProviderResources hook to be async
* Modify addResources to allow async calls
* Modify getApiKeyResources to handle key repairs when enabled
* Modify getApiKeyOutputs to use the resolvedLogicalApiKeyId resolved in getApiKeyResources